Fix encoding of completely empty tables.
Previously a table (which is not part of an array) has been ignored by the
encoder (except for its sub-tables) if it contained no key/value pairs.
Now the encoder creates the header of an empty table (with no key/value pairs
and no sub-tables) as this behavior is expected by the `toml-test` suit.
If a table is therefore not created implicitly through a sub-table, it
still appears in the encoded output.

The existing test has been corrected to reflect this change.