Add abstract syntax tree
Currently, there is no abstract representation of TOML in this
library. The parser uses `Map`s internally to represent objects
such as key/value pairs and table definitions. Also, the semantics
are tightly coupled with the parser. By introducing an abstract
syntax tree, I hope to simplify the parser implementation. The
AST could also be used by the encoder for pretty printing.