Allow subclasses with `const` constructor of encodable interface
In previous versions `TomlEncodable` was just an interface.
While its successors `TomlEncodableValue` and `TomlEncodableKey`
can still be used as an interface, the default implementation
of `TomlEncoableKey.toTomlKey` requires the `TomlEncodableKey` to be
extended or mixed in. Unlike interfaces, classes must have a
`const` constructor in order to permit `const` constructors of
sub classes.